BackEnd/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pyserial==3.5
